[CIT-394] Create unix/windows module for flockfile
